Novas atualizações

+++ products/views.py
@@ -4,13 +4,16 @@ from .serializers import CategorySerializer, ProductsSerializer
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.permissions import AllowAny
+from rest_framework.permissions import AllowAny, IsAuthenticated
 from .filters import CategoryFilter, ProductsFilter
 from rest_framework.generics import ListAPIView
 from rest_framework.filters import OrderingFilter, SearchFilter
 from django_filters.rest_framework import DjangoFilterBackend
 from .pagination import CategoryPagination, ProductsPagination
 from .cart import Cart
+from rest_framework.permissions import IsAdminUser
+from rest_framework.exceptions import PermissionDenied
+
 
 class CategoryListView(ListAPIView):
     serializer_class = CategorySerializer
@@ -20,10 +23,9 @@ class CategoryListView(ListAPIView):
     pagination_class = CategoryPagination
 
     def get(self, request, *args, **kwargs):
-        # Usa o queryset filtrado automaticamente pelo filterset
         categories = self.filter_queryset(self.get_queryset())
         
-        # Usar a paginação
+      
         page = self.paginate_queryset(categories)
         if page is not None:
             serializer = self.get_serializer(page, many=True, context={'request': request})
@@ -45,19 +47,36 @@ class CategoryDetailView(APIView):
         serializer = self.serializer_class(category, context={'request': request})
         return Response(serializer.data, status=status.HTTP_200_OK)
 
+    def post(self, request, *args, **kwargs):
+        # Apenas administradores podem criar categorias
+        if not request.user.is_staff:
+            raise PermissionDenied("Somente administradores podem criar categorias.")
+        serializer = self.serializer_class(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
     def put(self, request, slug, *args, **kwargs):
+        
+        if not request.user.is_staff:
+            raise PermissionDenied("Somente administradores podem editar categorias.")
         category = get_object_or_404(Category, slug=slug)
-        serializer = self.serializer_class(category)
+        serializer = self.serializer_class(category, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-  
     def delete(self, request, slug, *args, **kwargs):
+        # Apenas administradores podem deletar categorias
+        if not request.user.is_staff:
+            raise PermissionDenied("Somente administradores podem deletar categorias.")
         category = get_object_or_404(Category, slug=slug)
         category.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
 
 class ProductsListView(ListAPIView):
     serializer_class = ProductsSerializer
@@ -67,7 +86,7 @@ class ProductsListView(ListAPIView):
     ordering_fields = ['name', 'price', 'category'] 
     ordering = ['name']  
     search_fields = ['name', 'description']
-    pagination_class = ProductsPagination  # Classe de paginação
+    pagination_class = ProductsPagination  
 
     def get(self, request, *args, **kwargs):        
         products = self.filter_queryset(self.get_queryset())       
@@ -82,6 +101,8 @@ class ProductsListView(ListAPIView):
         return Products.objects.all()
 
 
+
+
 class ProductDetailView(APIView):
     serializer_class = ProductsSerializer
     permission_classes = [AllowAny]
@@ -92,43 +113,41 @@ class ProductDetailView(APIView):
         serializer = self.serializer_class(product, context={'request': request})
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def post(self, request, category_slug, category_id, product_slug, product_id, *args, **kwargs):
-        # Adiciona ou atualiza um produto no carrinho
+    def post(self, request, category_slug, category_id, *args, **kwargs):
+        if not request.user.is_staff:
+            raise PermissionDenied("Somente administradores podem adicionar produtos.")
+        category = get_object_or_404(Category, slug=category_slug, id=category_id)
+        serializer = self.serializer_class(data=request.data)
+        if serializer.is_valid():
+            serializer.save(category=category)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def put(self, request, category_slug, category_id, product_slug, product_id, *args, **kwargs):
+        if not request.user.is_staff:
+            raise PermissionDenied("Somente administradores podem editar produtos.")
         category = get_object_or_404(Category, slug=category_slug, id=category_id)
         product = get_object_or_404(Products, slug=product_slug, id=product_id, category=category)
-        
-        # Inicializa o carrinho
-        cart = Cart(request)
-
-        # Adiciona o produto ao carrinho (quantidade fornecida pela requisição)
-        quantity = request.data.get('quantity', 1)
-        override_quantity = request.data.get('override_quantity', False)
-        cart.add(product=product, quantity=quantity, override_quantity=override_quantity)
-
-        return Response({'message': 'Product added/updated in cart'}, status=status.HTTP_200_OK)
+        serializer = self.serializer_class(product, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, category_slug, category_id, product_slug, product_id, *args, **kwargs):
-        # Remove um produto do carrinho
+        if not request.user.is_staff:
+            raise PermissionDenied("Somente administradores podem deletar produtos.")
         category = get_object_or_404(Category, slug=category_slug, id=category_id)
         product = get_object_or_404(Products, slug=product_slug, id=product_id, category=category)
-        
-        # Inicializa o carrinho
-        cart = Cart(request)
-
-        # Remove o produto do carrinho
-        cart.remove(product=product)
-
-        return Response({'message': 'Product removed from cart'}, status=status.HTTP_204_NO_CONTENT)
+        product.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class CartDetailView(APIView):
-    permission_classes = [AllowAny]
-
+    permission_classes = [IsAuthenticated]
+    
     def get(self, request, *args, **kwargs):
-        # Inicializa o carrinho
         cart = Cart(request)
-        
-        # Itera sobre os produtos no carrinho
         products = []
         for item in cart:
             product = item['product']
@@ -140,6 +159,37 @@ class CartDetailView(APIView):
                 'total_price': item['total_price']
             }
             products.append(product_data)
-        
-        # Retorna a lista de produtos com suas quantidades e preços totais
+
         return Response({'products': products, 'total_price': cart.get_total_price()}, status=200)
+
+
+class CartView(APIView):
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request, *args, **kwargs):
+        product_id = request.data.get('product_id')
+        quantity = request.data.get('quantity', 1)
+        override_quantity = request.data.get('override_quantity', False)
+
+        try:
+            product = Products.objects.get(id=product_id)
+        except Products.DoesNotExist:
+            return Response({"error": "Produto não encontrado"}, status=status.HTTP_404_NOT_FOUND)
+
+        cart = Cart(request)
+        cart.add(product=product, quantity=quantity, override_quantity=override_quantity)
+
+        return Response({'message': 'Produto adicionado/atualizado no carrinho'}, status=status.HTTP_200_OK)
+
+    def delete(self, request, *args, **kwargs):
+        product_id = request.data.get('product_id')
+
+        try:
+            product = Products.objects.get(id=product_id)
+        except Products.DoesNotExist:
+            return Response({"error": "Produto não encontrado"}, status=status.HTTP_404_NOT_FOUND)
+
+        cart = Cart(request)
+        cart.remove(product=product)
+
+        return Response({'message': 'Produto removido do carrinho'}, status=status.HTTP_200_OK)
